common/readable.c, common/object.c, common/map.c, include/libproto.h: Use size_t   instead of int for buffer sizes. common/readable.c, include/libproto.h: Make some functions not used elsewhere   static. common/readable.c: Get rid of sprintf().
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/server/trunk@11395 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ common/object.c
@@ -2309,7 +2309,7 @@ void replace_insert_ob_in_map(const char *arch_string, object *op) {
  * @todo
  * handle case orig_ob->nrof == 0 (meaning 1).
  */
-object *get_split_ob(object *orig_ob, uint32 nr, char *err, int size) {
+object *get_split_ob(object *orig_ob, uint32 nr, char *err, size_t size) {
     object *newob;
 
     if (orig_ob->nrof < nr) {
